chor: cleanup cli utils
